+++ CHANGELOG.md
@@ -46,6 +46,7 @@ _When adding new entries to the changelog, please include issue/PR numbers where
 ### Other changes in this release
 
  * Added `sno create-patch <refish>` - creates a JSON patch file, which can be applied using `sno apply` [#210](https://github.com/koordinates/sno/issues/210)
+ * Added `sno data ls` - shows a list of datasets in the sno repository [#203](https://github.com/koordinates/sno/issues/203)
  * `sno clone` now support shallow clones (`--depth N`) to avoid cloning a repo's entire history [#174](https://github.com/koordinates/sno/issues/174)
  * `sno log` now supports JSON output with `--output-format json` [#170](https://github.com/koordinates/sno/issues/170)
  * `sno meta get` now prints text items as text (not encoded as JSON) [#211](https://github.com/koordinates/sno/issues/211)